+++ Oqtane.Server/Controllers/PageController.cs
@@ -6,6 +6,7 @@ using Oqtane.Models;
 using Oqtane.Shared;
 using System.Linq;
 using Oqtane.Infrastructure;
+using Oqtane.Security;
 
 namespace Oqtane.Controllers
 {
@@ -13,11 +14,13 @@ namespace Oqtane.Controllers
     public class PageController : Controller
     {
         private readonly IPageRepository Pages;
+        private readonly IUserPermissions UserPermissions;
         private readonly ILogManager logger;
 
-        public PageController(IPageRepository Pages, ILogManager logger)
+        public PageController(IPageRepository Pages, IUserPermissions UserPermissions, ILogManager logger)
         {
             this.Pages = Pages;
+            this.UserPermissions = UserPermissions;
             this.logger = logger;
         }
 
@@ -35,19 +38,26 @@ namespace Oqtane.Controllers
             }
         }
 
-        // GET api/<controller>/5
+        // GET api/<controller>/5?userid=x
         [HttpGet("{id}")]
-        public Page Get(int id)
+        public Page Get(int id, string userid)
         {
-            return Pages.GetPage(id);
+            if (userid == "")
+            {
+                return Pages.GetPage(id);
+            }
+            else
+            {
+                return Pages.GetPage(id, int.Parse(userid));
+            }
         }
 
         // POST api/<controller>
         [HttpPost]
-        [Authorize(Roles = Constants.AdminRole)]
+        [Authorize(Roles = Constants.RegisteredRole)]
         public Page Post([FromBody] Page Page)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && UserPermissions.IsAuthorized(User, "Edit", Page.Permissions))
             {
                 Page = Pages.AddPage(Page);
                 logger.Log(LogLevel.Information, this, LogFunction.Create, "Page Added {Page}", Page);
@@ -57,10 +67,10 @@ namespace Oqtane.Controllers
 
         // PUT api/<controller>/5
         [HttpPut("{id}")]
-        [Authorize(Roles = Constants.AdminRole)]
+        [Authorize(Roles = Constants.RegisteredRole)]
         public Page Put(int id, [FromBody] Page Page)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && UserPermissions.IsAuthorized(User, "Page", Page.PageId, "Edit"))
             {
                 Page = Pages.UpdatePage(Page);
                 logger.Log(LogLevel.Information, this, LogFunction.Update, "Page Updated {Page}", Page);
@@ -68,32 +78,38 @@ namespace Oqtane.Controllers
             return Page;
         }
 
-        // PUT api/<controller>/?siteid=x&parentid=y
+        // PUT api/<controller>/?siteid=x&pageid=y&parentid=z
         [HttpPut]
-        [Authorize(Roles = Constants.AdminRole)]
-        public void Put(int siteid, int? parentid)
+        [Authorize(Roles = Constants.RegisteredRole)]
+        public void Put(int siteid, int pageid, int? parentid)
         {
-            int order = 1;
-            List<Page> pages = Pages.GetPages(siteid).ToList();
-            foreach (Page page in pages.Where(item => item.ParentId == parentid).OrderBy(item => item.Order))
+            if (UserPermissions.IsAuthorized(User, "Page", pageid, "Edit"))
             {
-                if (page.Order != order)
+                int order = 1;
+                List<Page> pages = Pages.GetPages(siteid).ToList();
+                foreach (Page page in pages.Where(item => item.ParentId == parentid).OrderBy(item => item.Order))
                 {
-                    page.Order = order;
-                    Pages.UpdatePage(page);
+                    if (page.Order != order)
+                    {
+                        page.Order = order;
+                        Pages.UpdatePage(page);
+                    }
+                    order += 2;
                 }
-                order += 2;
+                logger.Log(LogLevel.Information, this, LogFunction.Update, "Page Order Updated {SiteId} {PageId} {ParentId}", siteid, pageid, parentid);
             }
-            logger.Log(LogLevel.Information, this, LogFunction.Update, "Page Order Updated {SiteId} {ParentId}", siteid, parentid);
         }
 
         // DELETE api/<controller>/5
         [HttpDelete("{id}")]
-        [Authorize(Roles = Constants.AdminRole)]
+        [Authorize(Roles = Constants.RegisteredRole)]
         public void Delete(int id)
         {
-            Pages.DeletePage(id);
-            logger.Log(LogLevel.Information, this, LogFunction.Delete, "Page Deleted {PageId}", id);
+            if (UserPermissions.IsAuthorized(User, "Page", id, "Edit"))
+            {
+                Pages.DeletePage(id);
+                logger.Log(LogLevel.Information, this, LogFunction.Delete, "Page Deleted {PageId}", id);
+            }
         }
     }
 }
